Struct taxi::Game

pub struct Game { /* fields omitted */ }

The Game with accompanying state

Methods

impl Game

fn new(print: bool) -> Game

Initialize a new game state

fn world_size(&self) -> (usize, usize)

Returns size of the game world

fn has_won(&self) -> bool

Returns true if player has won the game

fn passenger_position(&self) -> (u32, u32)

Passenger position

fn goal_position(&self) -> (u32, u32)

Goal position

fn player_position(&self) -> (u32, u32)

Returns player position

fn passenger_picked_up(&self) -> bool

Returns true if passenger has been picked up

fn distance_to_goal(&self) -> f64

Returns a distance to a game goal

fn distance_to_passenger(&self) -> f64

Returns a distance to a passenger

fn enter_move(&mut self, dir: Dir, print: bool) -> bool

Enter move directly and optionally print the map Returns a bool signifying whether or not this move lead to winning the game

fn make_move(&mut self, dir: Dir)

Makes a move and mutates the board, doesn't return anything

fn print_map(&self)

Print the map on screen